Moved the get_client function to the utils and removed the duplicated/redunant set_env calls. Changed the user_defined Client and Uploader to pass the argparse.Namespace up the stack to ensure that the lambda client used respects the aws-profile flag
