+++ panther_analysis_tool/util.py
@@ -17,6 +17,9 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+import argparse
+import boto3
+import logging
 import os
 from importlib import util as import_util
 from pathlib import Path
@@ -53,3 +56,19 @@ def store_modules(path: str, body: str) -> None:
     Path(os.path.dirname(path)).mkdir(parents=True, exist_ok=True)
     with open(path, "w") as py_file:
         py_file.write(body)
+
+
+def get_client(args: argparse.Namespace, service: str) -> boto3.client:
+    # optionally set env variable for profile passed as argument
+    if args.aws_profile is not None:
+        logging.info("Using AWS profile: %s", args.aws_profile)
+        set_env("AWS_PROFILE", args.aws_profile)
+        sess = boto3.Session(profile_name=args.aws_profile)
+        client = sess.client(service)
+    else:
+        client = boto3.client(service)
+    return client
+
+
+def set_env(key: str, value: str) -> None:
+    os.environ[key] = value
